Update modality_crossref.py
This is a project given by CK12 organization with the goal of obtaining information about certain educational resource pages for content development. This program takes in an input of a number that is linked to a page (in this case, artifact ID) and scrapes APIs to get information about the page, mainly hyperlinked words with incorrect destinations to fix those broken hyperlinks.